+++ src/Command/CacheWarmupCommand.php
@@ -2,6 +2,7 @@
 
 namespace Gnugat\MicroFrameworkBundle\Command;
 
+use Gnugat\MicroFrameworkBundle\Console\ExitCode;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -45,5 +46,7 @@ HELP
         $this->cacheWarmer->warmup($this->cacheDir);
 
         $io->success('Cache warmed up');
+
+        return ExitCode::SUCCESS;
     }
 }