+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Upgraded support to Symfony 5
 * Upgraded support to PHP 7.2 (accordingly with Symfony 5)
 * Upgraded tests to PHPUnit 8
+* added `Gnugat\MicroFrameworkBundle\Console\ExitCode`
 
 > **BC breaks**:
 >